refactor(immich_selfie_timelapse)!: Change default port to 2293
The first "breaking change"! Yes this is small, it's mostly to check in a real environment that it all works. Also 5000 is a very nice port to have relegated to a tiny tool that won't get used often.

I promise there will not be a lot of breaking changes like this, again, this is mostly to test it all works