docs: Add fancy application counter to Documentation homepage.
Kinda like the application count in the readme, I wanted to have something eye-catching to make the homepage a little less plain.

My original goal was to make this 100% in CSS (have the count calculated in the frontmatter and set a CSS variable), but it seems the keyframes can't animate a `@property` to a variable. Unfortunately everything else worked, and that would've been super cool.

Unfortunately some JS is needed. Instead of animating between keyframes, we can just put a transition on the variable.

In order to keep a good experience for anyone who has JS disabled, the count starts without rendering any content, then as soon as the window loads, a class is added that renders "0" and animates up. Otherwise no-JS people would just see "Browse all 0 applications," which definitely is not desired